requirements.txt
trio-chrome-devtools-protocol==0.6.0
requests
